# oapi/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Abandoned Factory Todo Lists
  servers:
    - url: http://localhost:8080/
paths:
  /chores:
    get:
      tags:
        - chores
      summary: Get list of all chores for a user
      operationId: getChores
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListChoresResponse'
  /chores/{choreId}/completed:
    put:
      tags:
        - chores
      summary: Mark this chore as having been completed just now
      operationId: markChoreCompleted
      parameters:
        - name: choreId
          in: path
          description: ID of chore completed
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chore'

  /lists:
    get:
      tags:
        - lists
      summary: Get list of Todo Lists for a user
      operationId: getLists
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetListsResponse'
components:
  schemas:
    Chore:
      type: object
      required:
        - id
        - name
        - last_completed
        - next_deadline
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        last_completed:
          type: integer
          format: int64
        next_deadline:
          type: integer
          format: int64
    ListChoresResponse:
      type: object
      required:
        - chores
      properties:
        chores:
          type: array
          items:
            $ref: '#/components/schemas/Chore'

    ListDescription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    GetListsResponse:
      type: object
      properties:
        lists:
          type: array
          items:
            $ref: '#/components/schemas/ListDescription'